pyproject.toml
[tool.poetry]
name = "py_nl2sql"
version = "0.0.1.dev1"
description = "A toolkit for converting natural language to SQL statements."
authors = ["liangzhu <pillarliang21@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/yourusername/your-package"
repository = "https://github.com/yourusername/your-package"

[tool.poetry.dependencies]
python = "^3.8.1"
numpy = "^1.21.0"
openai = "^1.43.0"
pydantic = "^2.8.2"
python-dotenv = "^1.0.1"
faiss-cpu = "^1.8.0.post1"
langchain-openai = "^0.1.23"
sqlalchemy = "^2.0.32"
pymysql = "^1.1.1"
poetry-core = "^1.9.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
